MNT: move TagsWidget and FlowLayout into own files
